+++ src/grasp/grafo.h
@@ -12,11 +12,11 @@ class Grafo
         int getCantidadVertices() const;
         
         // Los vertices se numeran de 0 a n-1 en los siguientes tres metodos
-        void setPesoArista(int u, int v, float weight);
-        float getPesoArista(int u, int v) const;
-        float getPesoAristasIncidentes(int u) const;
+        void setPesoArista(int u, int v, double weight);
+        double getPesoArista(int u, int v) const;
+        double getPesoAristasIncidentes(int u) const;
     private:
-        std::vector<std::vector<float>> matrizAdyacencia_;
+        std::vector<std::vector<double>> matrizAdyacencia_;
 };
 
 #endif /* GRAFO_H */
